pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "charity-django"
dynamic = ["version"]
description = "A set of Django apps for importing and using UK charity data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "David Kane", email = "david@dkane.net" },
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "dj_database_url",
    "Django >= 4.0",
    "faker",
    "inflect",
    "psycopg2-binary",
    "python-dateutil",
    "requests",
    "requests_cache",
    "requests_html",
    "titlecase",
    "tqdm",
]

[project.optional-dependencies]
test = [
    "black",
    "ruff",
    "pytest",
    "pytest-django",
    "pytest-pythonpath",
    "requests_mock",
    "testing.postgresql",
    "coverage",
]

[tool.hatch.envs.default]
features = [
    "test",
]

[tool.hatch.envs.default.scripts]
serve = "python test_project/manage.py runserver {args}"
test = "pytest src/"
test-cov = "coverage run -m pytest src/"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[project.urls]
Homepage = "https://github.com/kanedata/charity-django"

[tool.hatch.version]
path = "src/charity_django/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "src/charity_django",
]

[tool.ruff]
ignore = ['E501']